postgre - good luck

+++ messages.py
@@ -253,7 +253,7 @@ def polls(update: Update, context: CallbackContext):  # GROUP
         print(f"finally TIMESTAMP: {previous_timestamp}")
 
         if update.message.from_user.id in VERIFIED_USERS \
-                & int(previous_timestamp) + 20000 < now():  # 3628800000 < now():  ###enable again !!
+               and int(previous_timestamp) + 20000 < now():  # 3628800000 < now():  ###enable again !!
 
             update.message.delete()
